echo/requirements.txt
fastapi==0.104.1
mangum==0.17.0
xgboost==2.0.3
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2
joblib==1.3.2
google-generativeai==0.3.2
python-multipart==0.0.6
pydantic==2.5.3

